fix: change latitud and longitud types to Double in CrearDenunciaRequest; add Size validation to AdjuntarEvidenciaRequest

+++ ms-denuncias/src/main/java/com/andervalla/msdenuncias/controllers/dtos/requests/CrearDenunciaRequest.java
@@ -15,8 +15,8 @@ public record CrearDenunciaRequest(
         @NotBlank @Size(min = 5, max = 100) String titulo,
         @NotBlank @Size(min = 10, max = 500) String descripcion,
         @NotBlank CategoriaDenunciaEnum categoriaDenuncia,
-        @NotBlank @DecimalMin("-90.0") @DecimalMax("90.0") BigDecimal latitud,
-        @NotBlank @DecimalMin("-180.0") @DecimalMax("180.0") BigDecimal longitud,
+        @NotBlank @DecimalMin("-90.0") @DecimalMax("90.0") Double latitud,
+        @NotBlank @DecimalMin("-180.0") @DecimalMax("180.0") Double longitud,
         @NotBlank NivelAnonimatoEnum nivelAnonimato,
         @NotNull List<String> evidenciasIds
 ){}
